requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
yt-dlp>=2023.10.0
requests>=2.30.0
python-dotenv>=1.0.0
tqdm>=4.65.0
langdetect>=1.0.9
pysrt>=1.1.2
webvtt-py>=0.4.6
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
numpy>=1.24.0
isodate>=0.6.1